services/creative-generator-v2/requirements.txt
google-adk==1.28.0
fastapi
uvicorn
httpx
python-multipart
google-cloud-storage
google-auth
